% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clinical_events}
\alias{clinical_events}
\title{Time of clinical events}
\format{
A data frame with 2 rows and 2 variables:
\describe{
\item{Event}{type of event, ie. Progression or Death}
\item{Days_from_baseline}{time passed from first sampling time point, in days}
}
}
\usage{
clinical_events
}
\description{
A dataset containing the type and timing of clinical events.
}
\keyword{datasets}
