% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglongi.R
\name{gglongi}
\alias{gglongi}
\title{Title}
\usage{
gglongi(
  data,
  x_val,
  y_val,
  col_val,
  col_legend = T,
  mean_line = T,
  x_title = x_val,
  y_title = y_val,
  treatments = NA,
  scans = NA,
  progression = NA,
  treatment_start,
  treatment_end,
  treatment_drug,
  scan_date,
  event_date
)
}
\arguments{
\item{event_date}{}
}
\value{

}
\description{
Title
}
