% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mafs}
\alias{mafs}
\title{MAFs for 21 genes and 5 time points}
\format{
A data frame with 101 rows and 3 variables:
\describe{
\item{Gene}{gene symbol}
\item{time_from_baseline}{time passed from first sampling time point, in days}
\item{MAF}{mutant allele frequency}
}
}
\usage{
mafs
}
\description{
A dataset containing 5 longitudinal sampling time points and the associated MAF values for 21 genes.
}
\keyword{datasets}
