% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_val_per_timepoint.R
\name{mean_val_per_timepoint}
\alias{mean_val_per_timepoint}
\title{Calculate the mean value per time point}
\usage{
mean_val_per_timepoint(data, time_col, val_col)
}
\arguments{
\item{data}{The \code{data.frame} to be used for the calculation.

It should contain a column with the values to be summarized, and the corresponding time points denoting when the measurement was taken.

Zero-value entries are ignored and are completely removed from the calculation, unless every value corresponding to the time point is zero.}

\item{time_col}{\code{string} Name of the column containing the time data.}

\item{val_col}{\code{string} Name of the column containing the values to summarize.}
}
\value{
A \code{tibble} containing the mean value for each individual time point. The data frame contains two columns - the original \code{time_col} column and the calculated mean value column in the format \code{Mean_[val_col]}.
}
\description{
Calculate the mean value per time point
}
\examples{
x <- data.frame(
         days_from_baseline = c(0, 0, 0, 27, 27),
         MAF = c(0, 0.009, 0.007, 0.012, 0.032))

mean_val_per_timepoint(x, 'days_from_baseline', 'MAF')

}
