% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scans}
\alias{scans}
\title{Time of CT scans}
\format{
A data frame with 3 rows and 2 variables:
\describe{
\item{Scan_num}{order of the scan}
\item{Days_from_baseline}{time passed from first sampling time point, in days}
}
}
\usage{
scans
}
\description{
A dataset containing the times of performed CT scans.
}
\keyword{datasets}
