% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treatment}
\alias{treatment}
\title{Treatment data}
\format{
A data frame with 2 rows and 3 variables:
\describe{
\item{Treatment_drug}{name of the drug}
\item{start_from_baseline}{treatment start from first sampling time point, in days}
\item{stop_from_baseline}{treatment end from first sampling time point, in days}
}
}
\usage{
treatment
}
\description{
A dataset containing the time frames of treatment and the drugs used.
}
\keyword{datasets}
