% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglongi.R
\name{gglongi}
\alias{gglongi}
\title{Visualize longitudinal data}
\usage{
gglongi(
  data,
  x_val,
  y_val,
  col_val,
  col_legend = T,
  mean_line = T,
  x_title = x_val,
  y_title = y_val,
  treatments = NA,
  scans = NA,
  events = NA,
  treatment_start,
  treatment_end,
  treatment_drug,
  scan_date,
  event_date
)
}
\arguments{
\item{data}{The \code{data.frame} containing the data to be used for the visualization}

\item{x_val}{A \code{string} referring to the name of the column to be visualized on the X axis.}

\item{y_val}{A \code{string} referring to the name of the column to be visualized on the Y axis.}

\item{col_val}{A \code{string} referring to the name of the column to be used to color the points on the plot.}

\item{col_legend}{A \code{boolean} value. If \code{TRUE}, the color legend will be displayed.}

\item{mean_line}{A \code{boolean} value. If \code{TRUE}, the mean values per time point will be calculated and added as a continuous line.}

\item{x_title}{A \code{string} to be used as label for the X axis.}

\item{y_title}{A \code{string} to be used as label for the X axis.}

\item{treatments}{A \code{data.frame} containing treatment information. It should have 3 columns:
\itemize{
\item a column containing the name of the drug
\item a column containing the time information for the start of treatment
\item a column containing the time information for the end of treatment
}}

\item{scans}{A \code{data.frame} containing information about scan dates. It should contain a column denoting the time for each scan.}

\item{events}{A \code{data.frame} containing information on clinical events. It should contain the type of event and the associated timing for the event.}

\item{treatment_start}{A \code{string} referring to the name of the column containing the treatment start data.}

\item{treatment_end}{A \code{string} referring to the name of the column containing the treatment end data.}

\item{treatment_drug}{A \code{string} referring to the name of the column containing the treatment drug.}

\item{scan_date}{A \code{string} referring to the name of the column containing the scan date in the \code{scans} data frame.}

\item{event_date}{A \code{string} referring to the name of the column containing the event date in the \code{events} data frame.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualize longitudinal data
}
\examples{
mafs <- data.frame(time_from_baseline = c(0, 0, 0, 27, 27),
                   MAF = c(0, 0.009, 0.007, 0.012, 0.032),
                   Gene = c('BRAF', 'BRAF', 'BRAF', 'TP53', 'TP53'))

treatment = data.frame(Treatment_drug = c('Drug 1', 'Drug 2'),
                       start_from_baseline = c(0, 307),
                       stop_from_baseline = c(224, 538))

scans = data.frame(Scan_num = c(1, 2, 3),
                   Days_from_baseline = c(12, 247, 499))

clinical_events = data.frame(Event = c('Progression', 'Death'),
                             Days_from_baseline = c(273, 620))

gglongi(mafs, x_val = "time_from_baseline", y_val = "MAF", col_val = "Gene",
        col_legend = T, mean_line = T, x_title = "Days from baseline sample",
        y_title = "Mutant allele frequency", treatments = treatment, scans = scans,
        events = clinical_events, treatment_start = "start_from_baseline",
        treatment_end = "stop_from_baseline", treatment_drug = "Treatment_drug",
        scan_date = "Days_from_baseline", event_date = "Days_from_baseline")

}
