% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_val_per_timepoint.R
\name{mean_val_per_timepoint}
\alias{mean_val_per_timepoint}
\title{Calculate the mean value per time point}
\usage{
mean_val_per_timepoint(data, time_col, val_col)
}
\arguments{
\item{data}{A data frame containing the time points and values.}

\item{time_col}{Name of the column containing the time points.}

\item{val_col}{Name of the column containing the values to summarize.}
}
\value{
A data frame
}
\description{
Calculate the mean value per time point
}
